mvs-text2Answer/requirements.txt
langchain_aws
langgraph
python-dotenv==0.21.0
boto3==1.35.93



# pip install langchain langgraph boto3 botocore




